Reformat code and get rid of JsonStreamingParser subdirectory
