--- src/JsonStreamingParser/ParsingError.php
@@ -1,12 +1,0 @@
-<?php namespace JsonStreamingParser;
-
-class ParsingError extends \Exception {
-  /**
-   * @param int $line
-   * @param int $char
-   * @param string $message
-   */
-  public function __construct($line, $char, $message) {
-    parent::__construct("Parsing error in [$line:$char]. " . $message);
-  }
-}
