+++ src/Listener.php
@@ -1,0 +1,25 @@
+<?php
+namespace JsonStreamingParser;
+
+interface Listener
+{
+    public function start_document();
+
+    public function end_document();
+
+    public function start_object();
+
+    public function end_object();
+
+    public function start_array();
+
+    public function end_array();
+
+    // Key will always be a string
+    public function key($key);
+
+    // Note that value may be a string, integer, boolean, etc.
+    public function value($value);
+
+    public function whitespace($whitespace);
+}
